requirements.txt
arena
cv2
pillow
beautifulsoup4
numpy
scikit-image
